requirements.txt
efficientnet_pytorch==0.7.1
numpy==1.21.5
pandas==1.4.2
pretrainedmodels==0.7.4
scikit_learn==1.0.2
torch==1.11.0
torchvision==0.12.0
tensorboardx==2.2
pillow==9.0.1
